Check for potential taxonomic splits in a query

Check for potential taxonomic splits in a query

reptSplitCheck(
  x,
  pubDate = NULL,
  includeAll = FALSE,
  verbose = TRUE,
  cores = 1,
  showProgress = TRUE,
  exact = FALSE
)

Arguments

x

A character vector of species names to check. Usually from a database.

pubDate

Integer. An year (e.g., 2019) used as a reference date from when to check potential taxonomic split

includeAll

Logical; If TRUE, include all species described since pubDate regardless of if it is already included in the queried species list. Default is FALSE

verbose

Logical; If TRUE, prints progress messages. Default is TRUE.

cores

Integer. Number of CPU cores to use for parallel processing. Default is cores = 1.

showProgress

Logical. If TRUE, prints data sampling progress. Default is TRUE.

exact

Logical. Will search queried names for exact matches only (e.g., does not retrieve "Tantilla cf. melanocephala" when searching for "Tantilla melanocephala"). Default is FALSE.

Value

A data frame with the following columns:

query: the original input names.

RDB: the best-matching valid names according to The Reptile Database.

status: a status label indicating the result of the match ("check_split", "up_to_date", "not_found", or "failed").

Examples

query <- c(
"Atractus dapsilis",
"Atractus trefauti",
"Atractus snethlageae",
"Tantilla melanocephala",
"Oxybelis aeneus",
"Oxybelis rutherfordi",
"Vieira-Alencar authoristicus",
"Oxybelis aeneus",
"Bothrops pauloensis")

# \donttest{
result <- reptSplitCheck(x=query,
                        pubDate = 2019,
                        cores = 2,
                        showProgress = FALSE)

result <- reptSplitCheck(x=query,
                        pubDate = 2019,
                        cores = 2,
                        showProgress = FALSE,
                        includeAll = TRUE)
# }