Retrieve Synonyms for Reptile Species from RDB

Retrieves a data frame containing the current valid names of reptile species along with all their recognized synonyms, as listed in The Reptile Database (RDB). Optionally, the references citing each synonym can also be included.

reptSynonyms(x, getRef = FALSE, showProgress = TRUE, cores = 1)

Arguments

x

A character string with a species binomial or a data frame with columns species and url, typically the output of reptSpecies with getLink = TRUE.

getRef

Logical. If TRUE, includes the reference(s) in which each synonym was mentioned. Default is FALSE.

showProgress

Logical. If TRUE, prints data sampling progress. Default is TRUE.

cores

Integer. Number of CPU cores to use for parallel processing. Default is cores = 1.

Value

A data frame with columns:

species: The valid species name according to RDB.

synonym: Recognized synonyms and chresonyms for the species. Chresonyms are usually separated from authors with an emdash.

reference (optional): If getRef = TRUE, the citation where the synonym was reported.

References

Uetz, P., Freed, P., Aguilar, R., Reyes, F., Kudera, J., & Hošek, J. (eds.) (2025). The Reptile Database. Retrieved from http://www.reptile-database.org

See also

Examples

# Filter species belonging to genus Boa
boa <- letsRept::allReptiles[grep("^Boa\\s", letsRept::allReptiles$species), ]

# \donttest{
# Retrieve synonyms (without references)
boa_syn <- reptSynonyms(boa, getRef = FALSE, cores = 2)
#> 
#> Synonyms sampling complete!
Bconstrictor_syn <- reptSynonyms(x = "Boa constrictor")
#> 
#> Synonyms sampling complete for Boa constrictor 
#> 
#> Synonyms sampling complete!
# }